requirements.txt
python-binance
python-telegram-bot
pyyaml
pandas